Expose float values from trig_dist().

+++ src/line.h
@@ -77,8 +77,8 @@ std::vector<point> line_to( const point &p1, const point &p2, int t = 0 );
 // t and t2 decide which Bresenham line is used.
 std::vector<tripoint> line_to( const tripoint &loc1, const tripoint &loc2, int t = 0, int t2 = 0 );
 // sqrt(dX^2 + dY^2)
-int trig_dist(int x1, int y1, int x2, int y2);
-int trig_dist(const tripoint &loc1, const tripoint &loc2);
+float trig_dist(int x1, int y1, int x2, int y2);
+float trig_dist(const tripoint &loc1, const tripoint &loc2);
 // Roguelike distance; minimum of dX and dY
 int square_dist(int x1, int y1, int x2, int y2);
 int square_dist(const tripoint &loc1, const tripoint &loc2);
